+++ src/main/java/carpetfixes/mixins/advanced/ThreadedAnvilChunkStorage_chunkTimingsMixin.java
@@ -40,7 +40,7 @@ public class ThreadedAnvilChunkStorage_chunkTimingsMixin {
             constant = @Constant(longValue = 10000L)
     )
     private static long modifyChunkSavingCooldown(long original) {
-        return CFSettings.reIntroduceVeryAggressiveSaving ? -1L : CFSettings.chunkSaveCooldownDelay;
+        return CFSettings.reIntroduceVeryAggressiveSaving ? -1L : (long)CFSettings.chunkSaveCooldownDelay;
     }
